requirements.txt
arabic-reshaper==2.1.1
asgiref==3.3.1
backports.zoneinfo==0.2.1
certifi==2020.12.5
cffi==1.14.5
cryptography==3.4.7
cx-Oracle==8.1.0
Django==3.1.7
django-cors-headers==3.7.0
django-datetime-utc==1.0.3
django-ipware==3.0.2
django-reset-migrations==0.4.0
django-session-timeout==0.1.0
djangorestframework==3.12.2
et-xmlfile==1.0.1
fontawesome-free==5.15.3
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
jplephem==2.15
lunardate==0.2.0
numpy==1.20.2
openpyxl==3.0.7
Pillow==8.1.2
ply==3.11
psycopg2==2.8.6
pyCalverter==1.6.1
pycparser==2.20
PyJWT==1.7.1
pyluach==1.2.1
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2021.1
pyzk==0.9
reportlab==3.5.66
sgp4==2.18
six==1.15.0
skyfield==1.38
skyfield-data==3.0.0
sqlparse==0.4.1
thriftpy2==0.4.14
tzdata==2021.1
waitress==2.0.0
webencodings==0.5.1
workalendar==15.1.0
xhtml2pdf==0.2.5
xlwt==1.3.0
zk==0.6.10
